fix threading issues & simplify locking code
This commit makes sure all incoming image messages are handled in the
GUI thread. There were several calls in callbackImage() that called
Qt GUI methods on widgets in ui_, which are not thread-safe and can only
be used in the main thread.

By using Qt::QueuedConnection we can make sure that everything is done
in the main thread, which also simplifies the logic in RatioLayoutedFrame,
which - contrary to standard Qt practice - had its own locks to be
thread-safe.